+++ src/main/osd/boot.c
@@ -48,7 +48,6 @@
 #include "drivers/bus_spi.h"
 #include "drivers/flash_m25p16.h"
 #include "drivers/video_textscreen.h"
-#include "drivers/video_max7456.h"
 #include "drivers/usb_io.h"
 
 #include "fc/rc_controls.h" // FIXME for throttle status, not needed by OSD.
