+++ src/main/osd/cleanflight_osd.c
@@ -41,6 +41,7 @@
 #include "drivers/gpio.h"
 #include "drivers/system.h"
 #include "drivers/serial.h"
+#include "drivers/video_textscreen.h"
 
 #include "fc/rc_controls.h" // FIXME required due to virtual current meter.
